requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
python-dotenv==1.1.1
opencv-python==4.10.0.84
ultralytics==8.4.14
onnx==1.19.1
onnxruntime==1.21.1
